# Minor perf improvement to GEOSCoordSeq_copyToBuffer