--- a/benchmarks/capi/GEOSCoordSeqPerfTest.cpp
+++ b/benchmarks/capi/GEOSCoordSeqPerfTest.cpp
@@ -69,11 +69,60 @@ static void BM_CoordSeq_CreateByCoordinate(benchmark::State& state) {
             if (dim == 3) {
                 double z = *ptr++;
                 GEOSCoordSeq_setXYZ(seq, i, x, y, z);
+                benchmark::DoNotOptimize(z);
             }
+            benchmark::DoNotOptimize(x);
+            benchmark::DoNotOptimize(y);
         }
 
         GEOSCoordSeq_destroy(seq);
     }
+
+    finishGEOS();
+}
+
+template<size_t N, size_t dim>
+static void BM_CoordSeq_CopyByCoordinate(benchmark::State& state) {
+    initGEOS(nullptr, nullptr);
+
+    auto buf = create_buffer(N, dim);
+    auto seq = GEOSCoordSeq_copyFromBuffer(buf.data(), N, dim == 3, false);
+
+    for (auto _ : state) {
+        double x;
+        double y;
+        double z;
+        for (std::size_t i = 0; i < N; i++) {
+            if (dim == 2) {
+                GEOSCoordSeq_getXY(seq, i, &x, &y);
+            }
+            if (dim == 3) {
+                GEOSCoordSeq_getXYZ(seq, i, &x, &y, &z);
+                benchmark::DoNotOptimize(z);
+            }
+            benchmark::DoNotOptimize(x);
+            benchmark::DoNotOptimize(y);
+        }
+    }
+
+    GEOSCoordSeq_destroy(seq);
+
+    finishGEOS();
+}
+
+template<size_t N, size_t dim>
+static void BM_CoordSeq_CopyToBuffer(benchmark::State& state) {
+    initGEOS(nullptr, nullptr);
+
+    auto buf = create_buffer(N, dim);
+    auto seq = GEOSCoordSeq_copyFromBuffer(buf.data(), N, dim == 3, false);
+
+    for (auto _ : state) {
+        GEOSCoordSeq_copyToBuffer(seq, buf.data(), dim == 3, false);
+        benchmark::DoNotOptimize(buf);
+    }
+
+    GEOSCoordSeq_destroy(seq);
 
     finishGEOS();
 }
@@ -101,7 +150,13 @@ BENCHMARK_TEMPLATE(BM_CoordSeq_CreateByOrdinate, 10, 3);
 BENCHMARK_TEMPLATE(BM_CoordSeq_CreateByCoordinate, 10, 3);
 BENCHMARK_TEMPLATE(BM_CoordSeq_CopyFromBuffer, 10, 3);
 
-// N = 1,0000
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyByCoordinate, 10, 2);
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyToBuffer, 10, 2);
+
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyByCoordinate, 10, 3);
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyToBuffer, 10, 3);
+
+// N = 1,000
 BENCHMARK_TEMPLATE(BM_CoordSeq_CreateByOrdinate, 1000, 2);
 BENCHMARK_TEMPLATE(BM_CoordSeq_CreateByCoordinate, 1000, 2);
 BENCHMARK_TEMPLATE(BM_CoordSeq_CopyFromBuffer, 1000, 2);
@@ -109,6 +164,12 @@ BENCHMARK_TEMPLATE(BM_CoordSeq_CopyFromBuffer, 1000, 2);
 BENCHMARK_TEMPLATE(BM_CoordSeq_CreateByOrdinate, 1000, 3);
 BENCHMARK_TEMPLATE(BM_CoordSeq_CreateByCoordinate, 1000, 3);
 BENCHMARK_TEMPLATE(BM_CoordSeq_CopyFromBuffer, 1000, 3);
+
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyByCoordinate, 1000, 2);
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyToBuffer, 1000, 2);
+
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyByCoordinate, 1000, 3);
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyToBuffer, 1000, 3);
 
 // N = 10,000
 BENCHMARK_TEMPLATE(BM_CoordSeq_CreateByOrdinate, 10000, 2);
@@ -118,6 +179,12 @@ BENCHMARK_TEMPLATE(BM_CoordSeq_CopyFromBuffer, 10000, 2);
 BENCHMARK_TEMPLATE(BM_CoordSeq_CreateByOrdinate, 10000, 3);
 BENCHMARK_TEMPLATE(BM_CoordSeq_CreateByCoordinate, 10000, 3);
 BENCHMARK_TEMPLATE(BM_CoordSeq_CopyFromBuffer, 10000, 3);
+
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyByCoordinate, 10000, 2);
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyToBuffer, 10000, 2);
+
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyByCoordinate, 10000, 3);
+BENCHMARK_TEMPLATE(BM_CoordSeq_CopyToBuffer, 10000, 3);
 
 BENCHMARK_MAIN();