- Switch from std::memcpy to value copy
- Check for `CoordinateArraySequence` (likely) and so we can inline in that case

```
2022-02-06T12:43:43-05:00
Running bin/perf_capi_coordseq
Run on (8 X 3899.07 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 6144 KiB (x1)
Load Average: 1.96, 2.98, 2.84
-----------------------------------------------------------------------------------
Benchmark                                         Time             CPU   Iterations
-----------------------------------------------------------------------------------

BM_CoordSeq_CopyByCoordinate<10, 2>            30.0 ns         30.0 ns     23397094
BM_CoordSeq_CopyToBuffer<10, 2>                18.5 ns         18.5 ns     38914047
BM_CoordSeq_CopyByCoordinate<10, 3>            42.0 ns         42.0 ns     16709286
BM_CoordSeq_CopyToBuffer<10, 3>                16.6 ns         16.6 ns     41778813

BM_CoordSeq_CopyByCoordinate<1000, 2>          3121 ns         3121 ns       224390
BM_CoordSeq_CopyToBuffer<1000, 2>               540 ns          540 ns      1256236
BM_CoordSeq_CopyByCoordinate<1000, 3>          4148 ns         4147 ns       168217
BM_CoordSeq_CopyToBuffer<1000, 3>               777 ns          777 ns       885015

BM_CoordSeq_CopyByCoordinate<10000, 2>        31320 ns        31270 ns        22364
BM_CoordSeq_CopyToBuffer<10000, 2>             6652 ns         6643 ns       102111
BM_CoordSeq_CopyByCoordinate<10000, 3>        43295 ns        43293 ns        15862
BM_CoordSeq_CopyToBuffer<10000, 3>             8179 ns         8179 ns        84358
```